requirements.txt
Django~=2.0.6
psycopg2~=2.8.1
pillow~=6.0.0
django-import-export~=1.2.0